Modified a few of the text prompts with here docs. Did not modify all of them as I was unsatisfied with the solution. The extra credit asks that the printing of the here doc text be handled by the runner, which is not what this solution accomplishes.
I've been assuming there would be some method that would allow me to
print out the here doc string on any function, but I can't find it in
the docs or online.

The other issue is that if I am able to get the runner to handle that,
the runner should also be able to handle the user prompt and simply pass
the values returned by the user to the functions to be evaluated. This
suggests to me that the functions would need to take an argument, but
then how am I going to call the functions to print the hear docs before
the user has even been prompted?